+++ src/LocalStorage.js
@@ -1,3 +1,10 @@
+const {
+  GET_SNAPSHOT_TASK,
+  SET_SNAPSHOT_TASK,
+  CLEAR_SNAPSHOT_TASK,
+  NODE_EVENTS_INSTALLED,
+} = require("./constants");
+
 const LOCAL_STORAGE_METHODS = ["setItem", "getItem", "removeItem", "clear"];
 
 function logDisabled(method) {
@@ -23,17 +30,19 @@ class LocalStorage {
     ];
   }
 
-  constructor(localStorage, cy) {
+  constructor(localStorage, cy, Cypress) {
+    this._nodeEventsInstalled = Cypress.env(NODE_EVENTS_INSTALLED) === true;
+    this._snapshot = {};
     this._namedSnapshots = {};
     this._cy = cy;
     this._localStorage = localStorage;
+
     LOCAL_STORAGE_METHODS.forEach((localStorageMethod) => {
       this[logDisabledMethodName(localStorageMethod)] = logDisabled(localStorageMethod).bind(this);
     });
-    this.clearLocalStorageSnapshot();
   }
 
-  _saveLocalStorageKey(key, snapshotName) {
+  _saveLocalStorageKeyToMemory(key, snapshotName) {
     if (snapshotName) {
       this._namedSnapshots[snapshotName][key] = this._localStorage.getItem(key);
     } else {
@@ -41,7 +50,13 @@ class LocalStorage {
     }
   }
 
-  clearLocalStorageSnapshot(snapshotName) {
+  _saveLocalStorageToMemory(snapshotName) {
+    Object.keys(this._localStorage).forEach((key) => {
+      this._saveLocalStorageKeyToMemory(key, snapshotName);
+    });
+  }
+
+  _clearMemorySnapshot(snapshotName) {
     if (snapshotName) {
       this._namedSnapshots[snapshotName] = {};
     } else {
@@ -49,23 +64,61 @@ class LocalStorage {
     }
   }
 
+  _getSnapshotFromMemory(snapshotName) {
+    return snapshotName ? this._namedSnapshots[snapshotName] : this._snapshot;
+  }
+
+  _restoreLocalStorageFromSnapshot(obj = {}) {
+    Object.keys(obj).forEach((key) => {
+      this._localStorage.setItem(key, obj[key]);
+    });
+  }
+
+  _restoreLocalStorageFromMemory(snapshotName) {
+    this._restoreLocalStorageFromSnapshot(this._getSnapshotFromMemory(snapshotName));
+  }
+
+  _copySnapshotFromMemoryToNode(snapshotName) {
+    if (this._nodeEventsInstalled) {
+      return this._cy.task(SET_SNAPSHOT_TASK, {
+        name: snapshotName,
+        snapshot: this._getSnapshotFromMemory(snapshotName),
+      });
+    }
+  }
+
+  _clearNodeSnapshot(snapshotName) {
+    if (this._nodeEventsInstalled) {
+      return this._cy.task(CLEAR_SNAPSHOT_TASK, snapshotName);
+    }
+  }
+
+  _restoreLocalStorageFromNode(snapshotName) {
+    return this._cy.task(GET_SNAPSHOT_TASK, snapshotName).then((snapshot) => {
+      this._restoreLocalStorageFromSnapshot(snapshot);
+    });
+  }
+
+  clearLocalStorageSnapshot(snapshotName) {
+    this._clearMemorySnapshot(snapshotName);
+    return this._clearNodeSnapshot(snapshotName);
+  }
+
   saveLocalStorage(snapshotName) {
     if (!this._localStorage.getItem.wrappedMethod) {
       this.clearLocalStorageSnapshot(snapshotName);
-      Object.keys(this._localStorage).forEach((key) => {
-        this._saveLocalStorageKey(key, snapshotName);
-      });
+      this._saveLocalStorageToMemory(snapshotName);
+      return this._copySnapshotFromMemoryToNode(snapshotName);
     }
   }
 
   restoreLocalStorage(snapshotName) {
     this._localStorage.clear();
-    const snapshotToRestore = !!snapshotName
-      ? this._namedSnapshots[snapshotName] || {}
-      : this._snapshot;
-    Object.keys(snapshotToRestore).forEach((key) => {
-      this._localStorage.setItem(key, snapshotToRestore[key]);
-    });
+    if (this._nodeEventsInstalled) {
+      return this._restoreLocalStorageFromNode(snapshotName);
+    } else {
+      this._restoreLocalStorageFromMemory(snapshotName);
+    }
   }
 
   getLocalStorage(key) {