+++ src/constants.js
@@ -1,0 +1,11 @@
+const GET_SNAPSHOT_TASK = "cypressLocalStorageGetSnapshot";
+const SET_SNAPSHOT_TASK = "cypressLocalStorageSetSnapshot";
+const CLEAR_SNAPSHOT_TASK = "cypressLocalStorageClearSnapshot";
+const NODE_EVENTS_INSTALLED = "LOCALSTORAGE_NODE_EVENTS_INSTALLED";
+
+module.exports = {
+  GET_SNAPSHOT_TASK,
+  SET_SNAPSHOT_TASK,
+  CLEAR_SNAPSHOT_TASK,
+  NODE_EVENTS_INSTALLED,
+};