+++ test-e2e/cypress-10-no-plugin/src/modules/cookies-value/CookiesValue.js
@@ -1,0 +1,6 @@
+import { withData } from "@data-provider/react";
+
+import CookiesValue from "../../components/cookies-value";
+import { cookiesAccepted } from "../../data/user-preferences";
+
+export default withData(cookiesAccepted, "value")(CookiesValue);