+++ test-e2e/cypress-9/cypress/plugins/index.js
@@ -1,0 +1,11 @@
+const webpackPreprocessor = require("@cypress/webpack-preprocessor");
+const defaults = webpackPreprocessor.defaultOptions;
+
+const plugin = require("../../../../plugin");
+
+module.exports = (on, config) => {
+  plugin(on, config);
+  delete defaults.webpackOptions.module.rules[0].use[0].options.presets;
+  on("file:preprocessor", webpackPreprocessor(defaults));
+  return config;
+};